nvim: update config

+++ .config/nvim/utils/installer/install.sh
@@ -30,13 +30,24 @@ installnodefedora() {
     sudo dnf install -y npm
 }
 
+installnodegentoo() {
+	echo "Printing current node status..."
+	emerge -pqv net-libs/nodejs
+	echo "Make sure the npm USE flag is enabled for net-libs/nodejs"
+	echo "If it isn't enabled, would you like to enable it with flaggie? (Y/N)"
+	read answer
+        [ "$answer" != "${answer#[Yy]}" ] && sudo flaggie net-libs/nodejs +npm
+	sudo emerge -avnN net-libs/nodejs
+}
+
 installnode() {
 	echo "Installing node..."
 	[ "$(uname)" == "Darwin" ] && installnodemac
-	[ -n "$(uname -a | grep Ubuntu)" ] && installnodeubuntu
+	[ -n "$(cat /etc/os-release | grep Ubuntu)" ] && installnodeubuntu
 	[ -f "/etc/arch-release" ] && installnodearch
 	[ -f "/etc/artix-release" ] && installnodearch
 	[ -f "/etc/fedora-release" ] && installnodefedora
+	[ -f "/etc/gentoo-release" ] && installnodegentoo
 	[ "$(expr substr $(uname -s) 1 10)" == "MINGW64_NT" ] && echo "Windows not currently supported"
 	sudo npm i -g neovim
 }
@@ -52,25 +63,35 @@ installpiponubuntu() {
 }
 
 installpiponarch() {
-	pacman -S python-pip
+    sudo pacman -S python-pip
 }
 
 installpiponfedora() {
 	sudo dnf install -y pip >/dev/nul
 }
 
+installpipongentoo() {
+	sudo emerge -avn dev-python/pip
+}
+
 installpip() {
 	echo "Installing pip..."
 	[ "$(uname)" == "Darwin" ] && installpiponmac
-	[ -n "$(uname -a | grep Ubuntu)" ] && installpiponubuntu
+	[ -n "$(cat /etc/os-release | grep Ubuntu)" ] && installpiponubuntu
 	[ -f "/etc/arch-release" ] && installpiponarch
 	[ -f "/etc/fedora-release" ] && installpiponfedora
+	[ -f "/etc/gentoo-release" ] && installpipongentoo
 	[ "$(expr substr $(uname -s) 1 10)" == "MINGW64_NT" ] && echo "Windows not currently supported"
 }
 
 installpynvim() {
 	echo "Installing pynvim..."
-	pip3 install pynvim --user
+	if [ -f "/etc/gentoo-release" ]; then
+		echo "Installing using Portage"
+		sudo emerge -avn dev-python/pynvim
+	else
+		pip3 install pynvim --user
+	fi
 }
 
 installpacker() {
@@ -79,8 +100,8 @@ installpacker() {
 
 cloneconfig() {
 	echo "Cloning LunarVim configuration"
-	git clone https://github.com/ChristianChiarulli/lunarvim.git ~/.config/nvim
-	# mv $HOME/.config/nvim/init.lua $HOME/.config/nvim/init.lua.tmp
+	git clone --branch stable https://github.com/ChristianChiarulli/lunarvim.git ~/.config/nvim
+	mv $HOME/.config/nvim/utils/installer/lv-config.example.lua $HOME/.config/nvim/lv-config.lua
 	# mv $HOME/.config/nvim/utils/init.lua $HOME/.config/nvim/init.lua
 	nvim -u $HOME/.config/nvim/init.lua +PackerInstall
 	# rm $HOME/.config/nvim/init.lua
@@ -133,12 +154,19 @@ installonfedora() {
     pip3 install wheel ueberzug
 }
 
+installongentoo() {
+	sudo emerge -avn sys-apps/ripgrep app-shells/fzf app-misc/ranger dev-python/neovim-remote virtual/jpeg sys-libs/zlib
+	pipinstallueberzug
+	npm install -g tree-sitter-cli
+}
+
 installextrapackages() {
 	[ "$(uname)" == "Darwin" ] && installonmac
-	[ -n "$(uname -a | grep Ubuntu)" ] && installonubuntu
+	[ -n "$(cat /etc/os-release | grep Ubuntu)" ] && installonubuntu
 	[ -f "/etc/arch-release" ] && installonarch
 	[ -f "/etc/artix-release" ] && installonarch
 	[ -f "/etc/fedora-release" ] && installonfedora
+	[ -f "/etc/gentoo-release" ] && installongentoo
 	[ "$(expr substr $(uname -s) 1 10)" == "MINGW64_NT" ] && echo "Windows not currently supported"
 }
 
@@ -157,23 +185,23 @@ which node >/dev/null && echo "node installed, moving on..." || asktoinstallnode
 # install pynvim
 pip3 list | grep pynvim >/dev/null && echo "pynvim installed, moving on..." || installpynvim
 
-if [ -a "$HOME/.local/share/nvim/site/pack/packer/start/packer.nvim" ]; then
+if [ -e "$HOME/.local/share/nvim/site/pack/packer/start/packer.nvim" ]; then
 	echo 'packer already installed'
 else
 	installpacker
 fi
 
-if [ -a "$HOME/.config/nvim/init.lua" ]; then
-	echo 'nvcode already installed'
+if [ -e "$HOME/.config/nvim/init.lua" ]; then
+	echo 'LunarVim already installed'
 else
 	# clone config down
 	cloneconfig
 	# echo 'export PATH=$HOME/.config/nvim/utils/bin:$PATH' >>~/.zshrc
-	# echo 'export PATH=$HOME/.config/nvcode/utils/bin:$PATH' >>~/.bashrc
+	# echo 'export PATH=$HOME/.config/lunarvim/utils/bin:$PATH' >>~/.bashrc
 fi
 
 echo "I recommend you also install and activate a font from here: https://github.com/ryanoasis/nerd-fonts"
 
 # echo "I also recommend you add 'set preview_images_method ueberzug' to ~/.config/ranger/rc.conf"
 
-# echo 'export PATH=/home/$USER/.config/nvcode/utils/bin:$PATH appending to zshrc/bashrc'
+# echo 'export PATH=/home/$USER/.config/lunarvim/utils/bin:$PATH appending to zshrc/bashrc'
